トーラスアイ

+++ VioletVreath/src/jp/gecchi/VioletVreath/util/MyStgUtil.cpp
@@ -16,7 +16,7 @@ using namespace VioletVreath;
 		p->set(STAT_AddRankPoint, 0.00000 );  //‰ÁZƒ‰ƒ“ƒN
 		p->set(STAT_AddRankPoint_Reduction, 0.00000 );  //‰ÁZƒ‰ƒ“ƒNŒ¸­—¦
 		p->set(STAT_Stamina, 100 );  //‘Ì—Í
-		p->set(STAT_Attack, 105+((MyStraightLaserChip001::tex_no_*0.5)*100));  //UŒ‚—Í
+		p->set(STAT_Attack, (int)(105+((MyStraightLaserChip001::tex_no_*0.5)*100)));  //UŒ‚—Í
 		p->set(STAT_DefaultDefenceRate, 1.00000 );  //Šî€–hŒä—¦
 		p->set(STAT_DominantDefenceRate, 0.50000 );  //—D«‚Ì–hŒä—¦
 		p->set(STAT_RecessiveDefenceRate, 2.00000 );  //—ò«‚Ì–hŒä—¦
